Tests/LibWeb/Text/expected/hit_testing/link-inside-inline-stacking-context.txt

own text: true
link clicked, target=link
